Rename ga4-asset-fetch to ga4-mp and update refs
