+++ netlify/edge-functions/asset-tracking/index.test.ts
@@ -13,7 +13,7 @@ import test from 'node:test';
 import {
   ASSET_FETCH_GA_INFO_HEADER,
   INTERNAL_ASSET_FETCH_GA_INFO_VALUE,
-} from '../lib/ga4-asset-fetch.ts';
+} from '../lib/ga4-mp.ts';
 import {
   assertAssetFetchGa4Event,
   createWaitUntilSpy,
