+++ netlify/edge-functions/asset-tracking/live-check.test.mjs
@@ -12,7 +12,7 @@ import test from 'node:test';
 import {
   ASSET_FETCH_GA_INFO_HEADER,
   INTERNAL_ASSET_FETCH_GA_INFO_VALUE,
-} from '../lib/ga4-asset-fetch.ts';
+} from '../lib/ga4-mp.ts';
 import {
   absUrl,
   baseRef,
